+++ php/classes/FileAccess.php
@@ -1,8 +1,22 @@
 <?php
 
-class FileAccess {
-    public static function read($filepath, $waitLock = false, $default = null) {
-        $fileObj = array('filepath' => $filepath, 'content' => '');
+// strongly-typed file object
+class FileObject {
+    /** File path */
+    public string $filepath;
+    /** File content as string or JSON */
+    public string | array $content = '';
+    /** OS file descriptor (resource) */
+    public $fd;
+    public function __construct(string $filepath) {
+        $this->filepath = $filepath;
+    }
+}
+
+// basically a namespace
+abstract class FileAccess {
+    public static function read(string $filepath, $waitLock = false, $default = null) {
+        $fileObj = new FileObject($filepath);
         // open file as binary
         $file = fopen($filepath, 'rb');
 
@@ -11,17 +25,13 @@ class FileAccess {
             if ($default == null)
                 throw new Exception("Could not open file: $filepath");
 
-            // set default value
-            $fileObj['content'] = $default;
-        }
-
-        // if no file, puts default value inside
-        if ($file === false) {
-            file_put_contents($fileObj['filepath'], $fileObj['content'], LOCK_EX);
+            // set default value in file and try opening again
+            $fileObj->content = $default;
+            file_put_contents($fileObj->filepath, $fileObj->content, LOCK_EX);
             $file = fopen($filepath, 'rb');
         }
 
-        $fileObj['fd'] = $file;
+        $fileObj->fd = $file;
 
         // if want the lock, we wait for the shared lock
         if ($waitLock) {
@@ -38,7 +48,7 @@ class FileAccess {
             $string .= fread($file, 8192);
         }
 
-        $fileObj['content'] = $string;
+        $fileObj->content = $string;
 
         // if no wait you can close the file
         if (!$waitLock)
@@ -46,16 +56,16 @@ class FileAccess {
 
         return $fileObj;
     }
-    public static function write($fileObj) {
+    public static function write(FileObject $fileObj) {
         // lock and close
-        flock($fileObj['fd'], LOCK_UN);
-        fclose($fileObj['fd']);
+        flock($fileObj->fd, LOCK_UN);
+        fclose($fileObj->fd);
 
-        if (!is_writable($fileObj['filepath'])) {
+        if (!is_writable($fileObj->filepath)) {
             throw new HTTPException("PHP script can't write to file. Check permission, group and owner.", 400);
         }
 
-        $ret = file_put_contents($fileObj['filepath'], $fileObj['content'], LOCK_EX);
+        $ret = file_put_contents($fileObj->filepath, $fileObj->content, LOCK_EX);
         return $ret;
     }
 }