+++ php/files.php
@@ -135,7 +135,7 @@ if ($method === 'POST') {
         $imgInfo = getimagesize($absolutePath);
         if ($imgInfo === false) throw new Exception('Not an image');
 
-        header("Content-type: {$imgInfo['mime']}");
+        header("Content-Type: {$imgInfo['mime']}");
         header('Expires: 0');
         header('Cache-Control: must-revalidate, post-check=0, pre-check=0');
         header('Pragma: public');